Moved webLogic to server folder since it belongs there rather in the logic folder since its not business logic but UI logic
